copy structure over to sans2d
