+++ src/esssans/sans2d/masking.py
@@ -1,0 +1,88 @@
+# SPDX-License-Identifier: BSD-3-Clause
+# Copyright (c) 2023 Scipp contributors (https://github.com/scipp)
+"""
+"""
+from typing import NewType, Optional
+
+import scipp as sc
+
+from .types import (
+    CalibratedMaskedData,
+    CleanMasked,
+    DataWithLogicalDims,
+    DetectorEdgeMask,
+    MaskedData,
+    Numerator,
+    RunType,
+    SampleHolderMask,
+    SampleRun,
+)
+
+LowCountThreshold = NewType('LowCountThreshold', sc.Variable)
+"""Threshold below which detector pixels should be masked
+(low-counts on the edges of the detector panel, and the beam stop)"""
+
+
+def detector_edge_mask(sample: DataWithLogicalDims[SampleRun]) -> DetectorEdgeMask:
+    mask_edges = (
+        sc.abs(sample.coords['position'].fields.x) > sc.scalar(0.48, unit='m')
+    ) | (sc.abs(sample.coords['position'].fields.y) > sc.scalar(0.45, unit='m'))
+    return DetectorEdgeMask(mask_edges)
+
+
+def sample_holder_mask(
+    sample: DataWithLogicalDims[SampleRun],
+    low_counts_threshold: LowCountThreshold,
+) -> SampleHolderMask:
+    summed = sample.sum('tof')
+    holder_mask = (
+        (summed.data < low_counts_threshold)
+        & (sample.coords['position'].fields.x > sc.scalar(0, unit='m'))
+        & (sample.coords['position'].fields.x < sc.scalar(0.42, unit='m'))
+        & (sample.coords['position'].fields.y < sc.scalar(0.05, unit='m'))
+        & (sample.coords['position'].fields.y > sc.scalar(-0.15, unit='m'))
+    )
+    return SampleHolderMask(holder_mask)
+
+
+def mask_detectors(
+    da: DataWithLogicalDims[RunType],
+    edge_mask: Optional[DetectorEdgeMask],
+    holder_mask: Optional[SampleHolderMask],
+) -> MaskedData[RunType]:
+    """Apply pixel-specific masks to raw data.
+
+    Parameters
+    ----------
+    da:
+        Raw data.
+    edge_mask:
+        Mask for detector edges.
+    holder_mask:
+        Mask for sample holder.
+    """
+    da = da.copy(deep=False)
+    if edge_mask is not None:
+        da.masks['edges'] = edge_mask
+    if holder_mask is not None:
+        da.masks['holder_mask'] = holder_mask
+    return MaskedData[RunType](da)
+
+
+def mask_after_calibration(
+    da: CalibratedMaskedData[RunType],
+) -> CleanMasked[RunType, Numerator]:
+    return CleanMasked[RunType, Numerator](da)
+
+
+providers = (
+    detector_edge_mask,
+    sample_holder_mask,
+    mask_after_calibration,
+    mask_detectors,
+)
+"""
+Providers for loading and masking Sans2d data.
+
+These are meant for complementing the top-level :py:data:`esssans.providers` list.
+"""
