--- src/esssans/sans2d/sans2d.py
@@ -1,210 +1,0 @@
-# SPDX-License-Identifier: BSD-3-Clause
-# Copyright (c) 2023 Scipp contributors (https://github.com/scipp)
-"""
-Loading and masking specific to the ISIS Sans2d instrument and files stored in Scipp's
-HDF5 format.
-"""
-from typing import NewType, Optional
-
-import scipp as sc
-
-from .common import gravity_vector
-from .types import (
-    CalibratedMaskedData,
-    CleanMasked,
-    DataWithLogicalDims,
-    DetectorEdgeMask,
-    DetectorPixelShape,
-    DirectBeam,
-    DirectBeamFilename,
-    Filename,
-    LabFrameTransform,
-    LoadedFileContents,
-    MaskedData,
-    MonitorType,
-    NeXusMonitorName,
-    Numerator,
-    RawData,
-    RawMonitor,
-    RunNumber,
-    RunTitle,
-    RunType,
-    SampleHolderMask,
-    SampleRun,
-)
-
-LowCountThreshold = NewType('LowCountThreshold', sc.Variable)
-"""Threshold below which detector pixels should be masked
-(low-counts on the edges of the detector panel, and the beam stop)"""
-
-ReshapeToLogicalDims = NewType('ReshapeToLogicalDims', bool)
-"""Reshape raw data to logical dimensions if True"""
-
-
-def pooch_load(filename: Filename[RunType]) -> LoadedFileContents[RunType]:
-    from .data import get_path
-
-    dg = sc.io.load_hdf5(filename=get_path(filename))
-    data = dg['data']
-    if 'gravity' not in data.coords:
-        data.coords["gravity"] = gravity_vector()
-
-    # Some fixes specific for these Sans2d runs
-    sample_pos_z_offset = 0.053 * sc.units.m
-    # There is some uncertainty here
-    monitor4_pos_z_offset = -6.719 * sc.units.m
-
-    data.coords['sample_position'].fields.z += sample_pos_z_offset
-    # Results are actually slightly better at high-Q if we do not apply a bench offset
-    # bench_pos_y_offset = 0.001 * sc.units.m
-    # data.coords['position'].fields.y += bench_pos_y_offset
-    dg['monitors']['monitor4']['data'].coords[
-        'position'
-    ].fields.z += monitor4_pos_z_offset
-    return LoadedFileContents[RunType](dg)
-
-
-def pooch_load_direct_beam(filename: DirectBeamFilename) -> DirectBeam:
-    from .data import get_path
-
-    return DirectBeam(sc.io.load_hdf5(filename=get_path(filename)))
-
-
-def get_detector_data(
-    dg: LoadedFileContents[RunType],
-) -> RawData[RunType]:
-    return RawData[RunType](dg['data'])
-
-
-def get_monitor(
-    dg: LoadedFileContents[RunType], nexus_name: NeXusMonitorName[MonitorType]
-) -> RawMonitor[RunType, MonitorType]:
-    # See https://github.com/scipp/sciline/issues/52 why copy needed
-    mon = dg['monitors'][nexus_name]['data'].copy()
-    return RawMonitor[RunType, MonitorType](mon)
-
-
-def to_logical_dims(
-    da: RawData[RunType], reshape: Optional[ReshapeToLogicalDims]
-) -> DataWithLogicalDims[RunType]:
-    if reshape is None or not reshape:
-        return DataWithLogicalDims[RunType](da)
-    return DataWithLogicalDims[RunType](
-        da.fold(dim='spectrum', sizes={'y': -1, 'x': 1024})
-    )
-
-
-def detector_edge_mask(sample: DataWithLogicalDims[SampleRun]) -> DetectorEdgeMask:
-    mask_edges = (
-        sc.abs(sample.coords['position'].fields.x) > sc.scalar(0.48, unit='m')
-    ) | (sc.abs(sample.coords['position'].fields.y) > sc.scalar(0.45, unit='m'))
-    return DetectorEdgeMask(mask_edges)
-
-
-def sample_holder_mask(
-    sample: DataWithLogicalDims[SampleRun],
-    low_counts_threshold: LowCountThreshold,
-) -> SampleHolderMask:
-    summed = sample.sum('tof')
-    holder_mask = (
-        (summed.data < low_counts_threshold)
-        & (sample.coords['position'].fields.x > sc.scalar(0, unit='m'))
-        & (sample.coords['position'].fields.x < sc.scalar(0.42, unit='m'))
-        & (sample.coords['position'].fields.y < sc.scalar(0.05, unit='m'))
-        & (sample.coords['position'].fields.y > sc.scalar(-0.15, unit='m'))
-    )
-    return SampleHolderMask(holder_mask)
-
-
-def mask_detectors(
-    da: DataWithLogicalDims[RunType],
-    edge_mask: Optional[DetectorEdgeMask],
-    holder_mask: Optional[SampleHolderMask],
-) -> MaskedData[RunType]:
-    """Apply pixel-specific masks to raw data.
-
-    Parameters
-    ----------
-    da:
-        Raw data.
-    edge_mask:
-        Mask for detector edges.
-    holder_mask:
-        Mask for sample holder.
-    """
-    da = da.copy(deep=False)
-    if edge_mask is not None:
-        da.masks['edges'] = edge_mask
-    if holder_mask is not None:
-        da.masks['holder_mask'] = holder_mask
-    return MaskedData[RunType](da)
-
-
-def mask_after_calibration(
-    da: CalibratedMaskedData[RunType],
-) -> CleanMasked[RunType, Numerator]:
-    return CleanMasked[RunType, Numerator](da)
-
-
-def run_number(dg: LoadedFileContents[SampleRun]) -> RunNumber:
-    """Get the run number from the raw sample data."""
-    return RunNumber(int(dg['run_number']))
-
-
-def run_title(dg: LoadedFileContents[SampleRun]) -> RunTitle:
-    """Get the run title from the raw sample data."""
-    return RunTitle(dg['run_title'].value)
-
-
-def sans2d_tube_detector_pixel_shape() -> DetectorPixelShape[RunType]:
-    # Pixel radius and length
-    # found here:
-    # https://github.com/mantidproject/mantid/blob/main/instrument/SANS2D_Definition_Tubes.xml
-    R = 0.00405
-    L = 0.002033984375
-    pixel_shape = sc.DataGroup(
-        {
-            'vertices': sc.vectors(
-                dims=['vertex'],
-                values=[
-                    # Coordinates in pixel-local coordinate system
-                    # Bottom face center
-                    [0, 0, 0],
-                    # Bottom face edge
-                    [R, 0, 0],
-                    # Top face center
-                    [0, L, 0],
-                ],
-                unit='m',
-            ),
-            'nexus_class': 'NXcylindrical_geometry',
-        }
-    )
-    return pixel_shape
-
-
-def lab_frame_transform() -> LabFrameTransform[RunType]:
-    # Rotate +y to -x
-    return sc.spatial.rotation(value=[0, 0, 1 / 2**0.5, 1 / 2**0.5])
-
-
-providers = (
-    pooch_load_direct_beam,
-    pooch_load,
-    get_detector_data,
-    get_monitor,
-    detector_edge_mask,
-    sample_holder_mask,
-    mask_after_calibration,
-    mask_detectors,
-    run_number,
-    run_title,
-    to_logical_dims,
-    lab_frame_transform,
-    sans2d_tube_detector_pixel_shape,
-)
-"""
-Providers for loading and masking Sans2d data.
-
-These are meant for complementing the top-level :py:data:`esssans.providers` list.
-"""
